requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
et-xmlfile==1.1.0
exceptiongroup==1.0.0
h11==0.14.0
idna==3.4
ImageHash==4.3.1
numpy==1.23.4
openpyxl==3.0.10
outcome==1.2.0
pandas==1.5.1
Pillow==9.3.0
PySocks==1.7.1
pyTelegramBotAPI==4.7.1
python-dateutil==2.8.2
pytz==2022.6
PyWavelets==1.4.1
requests==2.28.1
scipy==1.9.3
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
wikipedia==1.4.0
wsproto==1.2.0
python-dotenv==0.21.0